% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\docType{data}
\name{CGC_20211118}
\alias{CGC_20211118}
\title{COSMIC-CGC genes mapped to PathwaySpace images.}
\format{
data.frame
}
\source{
COSMIC-CGC database (release v95, tier 1 collection).
}
\usage{
data(CGC_20211118)
}
\value{
A data.frame object.
}
\description{
A data frame listing 'GeneSymbol' and 'Entrez' IDs from the
COSMIC-CGC database (Sondka et al., 2020). These genes are used to
demonstrate the PathwaySpace's summit mapping pipeline, which assigns 
summits to an image space.
}
\examples{
data(CGC_20211118)
}
\references{
Sondka et al. The COSMIC Cancer Gene Census: describing genetic
dysfunction across all human cancers. Nat Rev Cancer 18, 696-705, 2018.
Doi: 10.1038/s41568-018-0060-1.
}
\keyword{CGC_20211118}
