% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-methods.R
\docType{methods}
\name{circularProjection,PathwaySpace-method}
\alias{circularProjection,PathwaySpace-method}
\alias{circularProjection}
\title{Circular Projection of Graph-Associated Signals}
\usage{
\S4method{circularProjection}{PathwaySpace}(
  ps,
  feature = activeFeature(ps),
  decay.fun = weibullDecay(),
  aggregate.fun = signalAggregation(),
  k = gs_vcount(ps),
  rescale = TRUE,
  verbose = TRUE,
  pdist = deprecated()
)
}
\arguments{
\item{ps}{A \linkS4class{PathwaySpace} class object.}

\item{feature}{A single string specifying the feature to project as a
signal. Must match either a feature name (see \code{gs_features(ps)}) or
a node attribute (see \code{gs_names(ps)}). If a node attribute, make sure
it is of numeric type. If the signal does not come from internal features,
assign it directly using the \code{\link{vertexSignal}} accessor.}

\item{decay.fun}{A signal decay function. Available options include
'Weibull', 'exponential', and 'linear' (see \code{\link{weibullDecay}}).
Users may also define a custom decay model with at least two arguments,
e.g., \code{function(x, signal) \{ ... \}}, which should returns a vector of
projected signals of the same length as \code{x}. Additional arguments may
include any variable available as a graph vertex attribute.}

\item{aggregate.fun}{A function used to aggregate the projected signals.
It must be provided as a unary function, e.g., \code{function(x) { ... }},
which should aggregate a vector of signals to a scalar value.
Available options include 'mean', 'wmean', 'log.wmean', and 'exp.wmean'
(See \code{\link{signalAggregation}}).}

\item{k}{A single positive integer specifying the maximum number of
vertices whose signals contribute to the projection. Defaults to
\code{gs_vcount(ps)}, i.e. all vertices are considered. Specifically,
at each point in space, the \emph{k}-top decayed signals are retained
prior to aggregation. Reducing \emph{k} focuses the projection on the
strongest local signals, filtering out weaker contributions.}

\item{rescale}{A logical value indicating whether to rescale
the signal. If the signal \code{>=0}, then it will be rescaled to
\code{[0, 1]}; if the signal \code{<=0}, then it will be rescaled to
\code{[-1, 0]}; and if the signal in \code{(-Inf, +Inf)}, then it will be
rescaled to \code{[-1, 1]}.}

\item{verbose}{A logical value specifying to display detailed
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}

\item{pdist}{Deprecated as of PathwaySpace 1.0.2; this parameter is now
passed internally through \code{decay.fun}.}
}
\value{
A preprocessed \linkS4class{PathwaySpace} class object.
}
\description{
\code{circularProjection()} implements a convolution algorithm to project
vertex-associated signals onto a 2D image space using a circular decay
function.
}
\examples{
library(PathwaySpace)

# Load a demo igraph
data('gtoy1', package = 'RGraphSpace')

# Create a new PathwaySpace object
ps <- buildPathwaySpace(gtoy1, nrc = 100)
# note: adjust 'nrc' to increase image resolution

# Set '1s' as vertex signal
vertexSignal(ps) <- 1

# Create a 2D-landscape image
ps <- circularProjection(ps)

}
\seealso{
\code{\link{buildPathwaySpace}}
}
\author{
Sysbiolab Team
}
