% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-misc.R
\name{getNearestNode}
\alias{getNearestNode}
\title{getNearestNode}
\usage{
getNearestNode(ps)
}
\arguments{
\item{ps}{Either a \linkS4class{PathwaySpace} or
\link[RGraphSpace]{GraphSpace} object.}
}
\value{
A \code{data.frame} with columns \code{from}, \code{to}, and
\code{dist}, listing each node's nearest neighbor and the Euclidean
distance between them.
}
\description{
Retrieves the nearest neighbor for each node from a
\linkS4class{PathwaySpace} object using Euclidean distance.
}
\examples{
# See examples in the PathwaySpace's tutorials:
# https://sysbiolab.github.io/PathwaySpace/

}
\seealso{
\code{\link[RANN]{nn2}}
}
