% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\docType{data}
\name{gimage}
\alias{gimage}
\title{An image matrix.}
\format{
matrix
}
\source{
This package.
}
\usage{
data(gimage)
}
\value{
An image matrix.
}
\description{
An image matrix used for workflow demonstrations.
}
\examples{
data(gimage)
}
\keyword{gimage}
