% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\docType{data}
\name{gtoys}
\alias{gtoys}
\alias{gtoy1}
\alias{gtoy2}
\alias{gtoy3}
\alias{gtoy3_1}
\alias{gtoy3_2}
\alias{gtoy3_3}
\title{Toy 'igraph' objects.}
\format{
igraph
}
\source{
This package.
}
\usage{
data(gtoy1)
}
\value{
Five igraph objects.
}
\description{
Small 'igraph' objects used for workflow demonstrations.
All graphs include 'coordX', 'coordY', and 'signal' vertex attributes.
}
\examples{
data(gtoy1)
data(gtoy2)
data(gtoy3)
}
\keyword{gtoys}
