% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\docType{data}
\name{Hallmarks_v2023_1_Hs_symbols}
\alias{Hallmarks_v2023_1_Hs_symbols}
\alias{hallmarks}
\title{A list with Hallmark gene sets (v2023.1)}
\format{
list
}
\source{
MSigDB database (v2023.1).
}
\usage{
data(Hallmarks_v2023_1_Hs_symbols)
}
\value{
A list object.
}
\description{
A list with Human gene symbols from the MSigDB's Hallmark gene set 
collection (Liberzon et al., 2015). These gene sets are used to
demonstrate the PathwaySpace's summit mapping pipeline, which assigns 
summits to an image space.
}
\examples{
data(Hallmarks_v2023_1_Hs_symbols)
}
\references{
Liberzon et al. The Molecular Signatures Database (MSigDB) hallmark 
gene set collection. Cell Systems 1(5):417-425, 2015
Doi: 10.1016/j.cels.2015.12.004
}
\keyword{Hallmarks_v2023_1_Hs_symbols}
