% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMethods.R
\name{length,PathwaySpace-method}
\alias{length,PathwaySpace-method}
\alias{length}
\title{Accessor function for PathwaySpace objects.}
\usage{
\S4method{length}{PathwaySpace}(x)
}
\arguments{
\item{x}{A \linkS4class{PathwaySpace} class object.}
}
\value{
A non-negative integer of length 1.
}
\description{
Get length of a PathwaySpace object.
}
\examples{
data('gtoy1', package = 'PathwaySpace')
pts <- buildPathwaySpace(gtoy1, nrc = 100)
length(pts)

}
