% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceDecay.R
\name{linearDecay}
\alias{linearDecay}
\title{A simple linear decay function}
\usage{
linearDecay(x, signal = 1, decay = 0.001)
}
\arguments{
\item{x}{A numeric vector of distances (in [0,1]).}

\item{signal}{A single numeric value representing a signal.
\code{\link{weibullDecay}} and \code{\link{expDecay}} functions.}

\item{decay}{A decay factor (in [0,1]). This term indicates how much the
\code{signal} decreases as a function of distance in pathway space. 
For example, at a specific distance defined by the \code{pdist} parameter 
(see \code{\link{circularProjection}}), the signal intensity will
be the initial signal multiplied by \code{decay}.}
}
\value{
A numeric vector; if missing 'x', it will return decay function.
}
\description{
The \code{linearDecay} function is used by PathwaySpace's
methods for signal convolution and projection.
}
\examples{
x <- seq(0, 2, 0.01)
y <- linearDecay(x, signal = 1)
plot(x, y)

}
\seealso{
\code{\link{weibullDecay}}, \code{\link{expDecay}}
}
\author{
Sysbiolab Team
}
