% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMethods.R
\docType{methods}
\name{names,PathwaySpace-method}
\alias{names,PathwaySpace-method}
\alias{names}
\title{Accessor functions for PathwaySpace objects.}
\usage{
\S4method{names}{PathwaySpace}(x)
}
\arguments{
\item{x}{A \linkS4class{PathwaySpace} class object.}
}
\value{
A character vector.
}
\description{
Get and set 'vertex' names of a \linkS4class{PathwaySpace}
class object.
}
\examples{
data('gtoy1', package = 'PathwaySpace')
pts <- buildPathwaySpace(gtoy1, nrc = 100)
names(pts)

}
