% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-classes.R
\docType{class}
\name{PathwaySpace-class}
\alias{PathwaySpace-class}
\title{PathwaySpace: An S4 class for signal projection on image spaces}
\value{
A \code{PathwaySpace} object.
}
\description{
\code{PathwaySpace} extends the \linkS4class[RGraphSpace]{GraphSpace} class with
signal projection slots. It stores projected signal matrices,
projection parameters, and workflow status, and is the main object used
by the \pkg{PathwaySpace} package.
}
\section{Slots}{

\describe{
\item{\code{projection}}{A \linkS4class{SpaceProjection} object storing the
intermediate and final matrices produced by a projection method.}

\item{\code{pars_ps}}{A list with PathwaySpace parameters.}

\item{\code{status}}{A vector containing the processing status of the
PathwaySpace object.}
}}

\section{Constructor}{

See \code{\link{buildPathwaySpace}}.
}

\seealso{
\linkS4class[RGraphSpace]{GraphSpace},
\linkS4class{SpaceProjection},
\code{\link{buildPathwaySpace}},
\code{\link{circularProjection}}
}
\author{
Sysbiolab Team, Mauro Castro (\email{mauro.castro@ufpr.br})
}
