% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\docType{package}
\name{PathwaySpace-package}
\alias{PathwaySpace}
\alias{PathwaySpace-package}
\title{PathwaySpace: Spatial Projection of Network Signals along Geodesic Paths}
\description{
For a given graph containing vertices, edges, and a signal associated with the vertices, the 'PathwaySpace' package performs a convolution operation, which involves a weighted combination of neighboring vertices and their associated signals. The package uses a decay function to project these signals, creating geodesic paths on a 2D-image space. 'PathwaySpace' has various applications, such as visualizing network data in a graphical format that highlights the relationships and signal strengths between vertices. By combining graph theory, signal processing, and visualization, 'PathwaySpace' provides a way of representing graph data on a continuous projection space. Based on methods introduced in Tercan et al. (2025) \doi{10.1016/j.xpro.2025.103681} and Ellrott et al. (2025) \doi{10.1016/j.ccell.2024.12.002}.
}
\section{Index}{

\tabular{ll}{
\link{PathwaySpace-class}:
\tab An S4 class for signal propagation on pathway spaces.\cr
\link{buildPathwaySpace}:
\tab Constructor of PathwaySpace-class objects.\cr
\link{circularProjection}:
\tab Creating 2D-landscape images from graph objects.\cr
\link{polarProjection}:
\tab Creating 2D-landscape images from graph objects.\cr
\link{silhouetteMapping}:
\tab Mapping graph silhouettes on PathwaySpace images.\cr
\link{summitMapping}:
\tab Mapping summits on a 2D-landscape image.\cr
\link{getPathwaySpace}:
\tab Accessory method for fetching slots from a PathwaySpace object.\cr
\link{plotPathwaySpace}:
\tab Plotting 2D-landscape images for the PathwaySpace package.\cr
}
Further information is available in the vignettes by typing
\code{vignette('PathwaySpace')}. Documented topics are also available in
HTML by typing \code{help.start()} and selecting the PathwaySpace package
from the menu.
}

\references{
Sysbiolab Team (2026). \emph{RGraphSpace: A lightweight interface between
'igraph' and 'ggplot2' graphics}. R package version 1.3.1 (Doi:
10.32614/CRAN.package.RGraphSpace),
\url{https://CRAN.R-project.org/package=RGraphSpace}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sysbiolab.github.io/PathwaySpace/}
  \item \url{https://github.com/sysbiolab/PathwaySpace}
  \item Report bugs at \url{https://github.com/sysbiolab/PathwaySpace/issues}
}

}
\author{
\strong{Maintainer}: Mauro Castro \email{mauro.a.castro@gmail.com} (\href{https://orcid.org/0000-0003-4942-8131}{ORCID})

Authors:
\itemize{
  \item Sysbiolab Team
}

Other contributors:
\itemize{
  \item Victor Apolonio [contributor]
  \item Jonathan Back [contributor]
  \item Lana Querne [contributor]
  \item Vinicius Chagas [contributor]
  \item Bahar Tercan [contributor]
}

}
\keyword{internal}
