% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceGraph.R
\docType{methods}
\name{plotGraphSpace,PathwaySpace-method}
\alias{plotGraphSpace,PathwaySpace-method}
\alias{plotGraphSpace}
\title{Plotting graphs for the PathwaySpace package.}
\usage{
\S4method{plotGraphSpace}{PathwaySpace}(
  pts,
  node.size = 1,
  edge.width = 0.5,
  node.color = "grey50",
  edge.color = "grey80",
  bg.color = "grey95",
  font.size = 1,
  theme.name = c("th0", "th1", "th2", "th3"),
  xlab = "Pathway coordinates 1",
  ylab = "Pathway coordinates 2",
  marks = FALSE,
  mark.size = 3,
  mark.color = "grey20"
)
}
\arguments{
\item{pts}{A \linkS4class{PathwaySpace} class object.}

\item{node.size}{A single numeric value for node size.}

\item{edge.width}{A single numeric value for edge width.}

\item{node.color}{A single color for node color.}

\item{edge.color}{A single color for edge color.}

\item{bg.color}{A single color for background.}

\item{font.size}{A single numeric value passed to ggplot themes.}

\item{theme.name}{Name of a custom PathwaySpace theme. These themes 
(from 'th0' to 'th3') consist mainly of preconfigured ggplot settings, 
which the user can subsequently fine-tune within the resulting 
ggplot object.}

\item{xlab}{The title for the 'x' axis of a 2D-image space.}

\item{ylab}{The title for the 'y' axis of a 2D-image space.}

\item{marks}{A logical value indicating whether to add 'marks' to vertex 
positions. Alternatively, this could be a vector listing vertex names.}

\item{mark.size}{A font size argument passed to \code{\link{geom_text}}.}

\item{mark.color}{A color passed to \code{\link{geom_text}}.}
}
\value{
A ggplot-class object.
}
\description{
\code{plotGraphSpace} is a wrapper function to 
create dedicated ggplot graphics for PathwaySpace-class objects.
}
\examples{
# Load a demo igraph
data('gtoy1', package = 'PathwaySpace')

# Create a new PathwaySpace object
pts <- buildPathwaySpace(gtoy1)

plotGraphSpace(pts)

}
\seealso{
\code{\link{circularProjection}}
}
