% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\name{plotPathDistances}
\alias{plotPathDistances}
\title{Accessory function to plot pathway space distances.}
\usage{
plotPathDistances(pdist, z.transform = FALSE)
}
\arguments{
\item{pdist}{A list generated by the \code{\link{pathDistances}} function.}

\item{z.transform}{A single logical value specifying to convert pathway 
distances into z-score values.}
}
\value{
A 'ggplot' object.
}
\description{
Accessory function to plot pathway space distances.
}
\examples{

data("PCv12_pruned_igraph", package = "PathwaySpace")
gdist <- igraph::distances(graph = PCv12_pruned_igraph, algorithm="unweighted")
from <- sample(colnames(gdist), 50)
to <- sample(colnames(gdist), 50)

# res <- pathDistances(gdist, from, to)
# plotPathDistances(res)

}
