% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-misc.R
\name{pspace.cols}
\alias{pspace.cols}
\title{A simple vector of colors for PathwaySpace images}
\usage{
pspace.cols(n = 25)
}
\arguments{
\item{n}{The number of colors to generate in the output palette.}
}
\value{
A vector with hexadecimal color codes.
}
\description{
A simple vector of colors for PathwaySpace images
}
\examples{
pspace.cols()

}
\seealso{
\code{\link{plotPathwaySpace}}, \code{\link{pspace.pals}}
}
