% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceDecay.R
\name{signalDecay}
\alias{signalDecay}
\title{Helper function to construct decay models}
\usage{
signalDecay(
  method = c("weibull", "exp", "linear"),
  decay = 0.001,
  shape = 1.05
)
}
\arguments{
\item{method}{A character string specifying a method for
signal decay (any of \code{weibull}, \code{exp}, or \code{linear}), 
returning \code{\link{weibullDecay}}, \code{\link{expDecay}}, 
or \code{\link{linearDecay}} functions, respectively.}

\item{decay}{A decay factor in the interval (0,1), passed to the 
\code{\link{weibullDecay}} and \code{\link{expDecay}} functions.}

\item{shape}{A parameter (>=1) passed to the \code{\link{weibullDecay}} 
function.}
}
\value{
Returns a function of the form: \code{function(x, signal) { ... }}
}
\description{
Utility functions to construct signal decay models used by
\code{\link{circularProjection}} and \code{\link{polarProjection}} 
internal calls.
}
\examples{
decay.fun <- signalDecay()

}
\seealso{
\code{\link{circularProjection}} and \code{\link{polarProjection}}
}
\author{
Sysbiolab Team
}
