% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-classes.R
\name{updatePathwaySpace,PathwaySpace-method}
\alias{updatePathwaySpace,PathwaySpace-method}
\alias{updatePathwaySpace}
\title{Update a PathwaySpace object}
\usage{
\S4method{updatePathwaySpace}{PathwaySpace}(x, verbose = TRUE)
}
\arguments{
\item{x}{A \code{PathwaySpace} object.}

\item{verbose}{Logical; if \code{TRUE}, reports which slots were added.}
}
\value{
An updated \code{PathwaySpace} object.
}
\description{
Updates outdated \code{PathwaySpace} objects serialized from
previous package versions, adding any missing slots with default values.
}
