% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMethods.R
\docType{methods}
\name{vertexSignal,PathwaySpace-method}
\alias{vertexSignal,PathwaySpace-method}
\alias{vertexSignal}
\alias{vertexSignal<-}
\alias{vertexSignal<-,PathwaySpace-method}
\title{Accessor functions for fetching slots from a PathwaySpace object.}
\usage{
\S4method{vertexSignal}{PathwaySpace}(ps)

\S4method{vertexSignal}{PathwaySpace}(ps) <- value
}
\arguments{
\item{ps}{A \linkS4class{PathwaySpace} class object.}

\item{value}{A numeric vector with values representing signal
intensities. This vector should be aligned to the "vertex" slot.}
}
\value{
A numeric vector.
}
\description{
Get or set 'signal' for a \linkS4class{PathwaySpace}
class object.
}
\examples{
data('gtoy1', package = 'RGraphSpace')
ps <- buildPathwaySpace(gtoy1, nrc = 100)

# Check vertex names
names(ps)

# Check vertex signal
vertexSignal(ps)

# Set signal to a given vertex
vertexSignal(ps)[1] <- 1

# Set signal to a given vertex
vertexSignal(ps)["n3"] <- 1

# Set '1s' to all vertices
vertexSignal(ps) <- 1

}
