% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\name{getNearestNode}
\alias{getNearestNode}
\title{getNearestNode}
\usage{
getNearestNode(ps)
}
\arguments{
\item{ps}{Either a \linkS4class{PathwaySpace} or 
\link[RGraphSpace]{GraphSpace} object.}
}
\description{
Retrieves the nearest neighbor for each node from a 
\linkS4class{PathwaySpace} object using Euclidean distance.
}
\examples{
# See examples in the PathwaySpace's tutorials:
# https://sysbiolab.github.io/PathwaySpace/

}
\seealso{
\code{\link[RANN]{nn2}}
}
