% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceClasses.R
\docType{class}
\name{PathwaySpace-class}
\alias{PathwaySpace-class}
\title{PathwaySpace: An S4 class for signal propagation on image spaces}
\value{
An S4 class object.
}
\description{
PathwaySpace: An S4 class for signal propagation on image spaces
}
\section{Slots}{

\describe{
\item{\code{nodes}}{A data frame with xy-vertex coordinates.}

\item{\code{edges}}{A data frame with edges.}

\item{\code{graph}}{An igraph object.}

\item{\code{image}}{A raster background image matrix.}

\item{\code{pars}}{A list with parameters.}

\item{\code{misc}}{A list with intermediate objects for downstream methods.}

\item{\code{projections}}{A list with processed objects for downstream methods.}

\item{\code{status}}{A vector containing the processing status of the
PathwaySpace object.}
}}

\section{Constructor}{

see \code{\link{buildPathwaySpace}} constructor.
}

\author{
Mauro Castro, \email{mauro.castro@ufpr.br}
}
