% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspace-misc.R
\docType{package}
\name{PathwaySpace-package}
\alias{PathwaySpace}
\alias{PathwaySpace-package}
\title{PathwaySpace: Spatial Projection of Network Signals along Geodesic Paths}
\description{
For a given graph containing vertices, edges, and a signal associated with the vertices, the 'PathwaySpace' package performs a convolution operation, which involves a weighted combination of neighboring vertices and their associated signals. The package uses a decay function to project these signals, creating geodesic paths on a 2D-image space. 'PathwaySpace' has various applications, such as visualizing network data in a graphical format that highlights the relationships and signal strengths between vertices. By combining graph theory, signal processing, and visualization, 'PathwaySpace' provides a way of representing graph data on a continuous projection space. Based on methods introduced in Tercan et al. (2025) \doi{10.1016/j.xpro.2025.103681} and Ellrott et al. (2025) \doi{10.1016/j.ccell.2024.12.002}.
}
\details{
For a hands-on introduction, see the vignette:
\code{vignette("PathwaySpace")}.

The full set of documented topics can also be browsed in HTML by
running \code{help.start()} and selecting the PathwaySpace package
from the package list.
}
\references{
Tercan B, Apolonio VH, Chagas VS, Wong CK, Lee JA, Yau C, Benz CC,
Stuart JM, Karlberg BJ, Ellrott K, Grewal JK, Jones SJ, Network TCGAA,
Zenklusen JC, Robertson AG, Laird PW, Cherniack AD, Castro MA (2025).
"Protocol for assessing distances in pathway space for classifier
feature sets from machine learning methods." \emph{STAR Protocols}, \emph{6}(2),
103681. doi:10.1016/j.xpro.2025.103681
\url{https://doi.org/10.1016/j.xpro.2025.103681}.

Ellrott K, Wong CK, Yau C, Castro MA, Lee JA, Karlberg BJ, Grewal JK,
Lagani V, Tercan B, al. e (2025). "Classification of non-TCGA cancer
samples to TCGA molecular subtypes using compact feature sets." \emph{Cancer
Cell}, \emph{43}(1), 1. doi:10.1016/j.ccell.2024.12.002
\url{https://doi.org/10.1016/j.ccell.2024.12.002}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sysbiolab.github.io/PathwaySpace/}
  \item \url{https://github.com/sysbiolab/PathwaySpace}
  \item Report bugs at \url{https://github.com/sysbiolab/PathwaySpace/issues}
}

}
\author{
\strong{Maintainer}: Mauro Castro \email{mauro.a.castro@gmail.com} (\href{https://orcid.org/0000-0003-4942-8131}{ORCID})

Authors:
\itemize{
  \item Sysbiolab Team
}

Other contributors:
\itemize{
  \item Victor Apolonio [contributor]
  \item Jonathan Back [contributor]
  \item Lana Querne [contributor]
  \item Vinicius Chagas [contributor]
  \item Bahar Tercan [contributor]
}

}
\keyword{internal}
