% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMethods.R
\docType{methods}
\name{polarProjection,PathwaySpace-method}
\alias{polarProjection,PathwaySpace-method}
\alias{polarProjection}
\title{Creating 2D-landscape images from graph objects.}
\usage{
\S4method{polarProjection}{PathwaySpace}(
  ps,
  k = 8,
  pdist = 0.5,
  theta = 180,
  directional = FALSE,
  rescale = TRUE,
  verbose = TRUE,
  decay.fun = signalDecay(),
  aggregate.fun = signalAggregation(),
  kns = deprecated(),
  knn = deprecated()
)
}
\arguments{
\item{ps}{A \linkS4class{PathwaySpace} class object.}

\item{k}{A single positive integer determining the k-top signals in the 
signal convolution operation.}

\item{pdist}{A term (in \code{[0, 1]}) determining a distance unit for the
signal decay function. When `pdist = 1` it will represent the diameter of 
the inscribed circle within the pathway space. This distance will affect the
extent over which the convolution operation projects the signal between
source- and destination points.}

\item{theta}{Angle of projection (degrees in \code{(0,360]}).}

\item{directional}{If directional edges are available, this argument can 
be used to orientate the signal projection on directed graphs.}

\item{rescale}{A single logical value indicating whether to rescale 
the signal. If the signal \code{>=0}, then it will be rescaled to 
\code{[0, 1]}; if the signal \code{<=0}, then it will be rescaled to 
\code{[-1, 0]}; and if the signal in \code{(-Inf, +Inf)}, then it will be 
rescaled to \code{[-1, 1]}.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}

\item{decay.fun}{A signal decay function. Available: 'Weibull', 
'exponential', and 'linear' functions (see \code{\link{signalDecay}}).}

\item{aggregate.fun}{A signal aggregation function. Available: 
weighted 'linear', 'log', and 'exponential' functions 
(see \code{\link{signalAggregation}}).}

\item{kns}{Deprecated from PathwaySpace 1.0.1; use 'k' instead.}

\item{knn}{Deprecated from PathwaySpace 1.0.1; use 'k' instead.}
}
\value{
A preprocessed \linkS4class{PathwaySpace} class object.
}
\description{
\code{polarProjection} implements a convolution algorithm
to project signals across a 2D-coordinate system.
}
\examples{
# Load a demo igraph
data('gtoy2', package = 'RGraphSpace')

# Create a new PathwaySpace object
ps <- buildPathwaySpace(gtoy2, nrc = 100)
# note: adjust 'nrc' to increase image resolution

# Set '1s' as vertex signal
vertexSignal(ps) <- 1

# Create a 2D-landscape image
ps <- polarProjection(ps)

}
\seealso{
\code{\link{buildPathwaySpace}}
}
\author{
Mauro Castro
}
