% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceMisc.R
\name{pspace.cols}
\alias{pspace.cols}
\title{A simple vector of colors for PathwaySpace images}
\usage{
pspace.cols(n = 5)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
A vector with hexadecimal color codes.
}
\description{
A simple vector of colors for PathwaySpace images
}
\examples{
pspace.cols()

}
\seealso{
\code{\link{plotPathwaySpace}}
}
