% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaceDecay.R
\name{signalDecay}
\alias{signalDecay}
\title{Deprecated function}
\usage{
signalDecay(...)
}
\arguments{
\item{...}{Deprecated arguments}
}
\value{
Stop unconditionally
}
\description{
Use \code{\link{weibullDecay}}, \code{\link{expDecay}}, 
and \code{\link{linearDecay}}.
}
\examples{
decay.fun <- weibullDecay()

}
\author{
Sysbiolab Team
}
